move config from share to AA
