+++ src/Authentication.Abstractions/Interfaces/IConfigManager.cs
@@ -16,7 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Management.Automation;
 
-namespace Microsoft.Azure.PowerShell.Common.Share.Config
+namespace Microsoft.Azure.PowerShell.Common.Config
 {
     public interface IConfigManager
     {
